fix(stack): fix accessibility issue
